requirement.txt
torch>=2.0
torchvision
tqdm
scikit-learn
pandas
Pillow
pyyaml
matplotlib
